requirements.txt
asgiref==3.10.0
Django==5.2.8
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
drf-yasg==1.21.11
pytz==2025.2
PyYAML==6.0.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0